adapt suggestion
