+++ src/lib/y2storage/boot_requirements_strategies/analyzer.rb
@@ -22,6 +22,9 @@ require "pathname"
 require "y2storage/planned"
 require "y2storage/encryption_type"
 
+Yast.import "ProductFeatures"
+Yast.import "Arch"
+
 module Y2Storage
   module BootRequirementsStrategies
     # Auxiliary class that takes information from several sources (current
@@ -372,6 +375,15 @@ module Y2Storage
       #   the planned devices or in the devicegraph) or is not encrypted.
       def encrypted_zipl?
         encrypted?(device_for_zipl)
+      end
+
+      def self.bls_bootloader_proposed?
+        preferred_bootloader = Yast::ProductFeatures.GetStringFeature("globals",
+          "preferred_bootloader")
+        Y2Storage::Arch.new.efiboot? &&
+          (Yast::Arch.x86_64 || Yast::Arch.aarch64) &&
+          !StorageEnv.instance.no_bls_bootloader &&
+          ["systemd-boot", "grub2-bls"].include?(preferred_bootloader)
       end
 
       protected
