+++ src/lib/y2storage/volume_specification.rb
@@ -411,7 +411,7 @@ module Y2Storage
     def adjust_features
       self.adjust_by_ram = false if swap? && !resume_supported?
 
-      return unless Y2Storage::BootRequirementsStrategies::BLS.bls_bootloader_proposed?
+      return unless Y2Storage::BootRequirementsStrategies::Analyzer.bls_bootloader_proposed?
 
       # Removing grub2/grub2-efi specific subvolumes because they are not needed.
       @subvolumes.delete_if do |subvol|
